+++ camera/__init__.py
@@ -1,0 +1,1 @@
+"Camera utilities and capture helpers."
