+++ camera/cv_grab_callback.py
@@ -1,7 +1,7 @@
 #coding=utf-8
 import cv2
 import numpy as np
-import mvsdk
+from . import mvsdk
 import platform
 import time
 import threading
