+++ camera/detection.json
