+++ camera/examples/Camera/cv_grab2.py
@@ -1,7 +1,7 @@
 #coding=utf-8
 import cv2
 import numpy as np
-import mvsdk
+from .. import mvsdk
 import platform
 
 class Camera(object):
