+++ camera/examples/Camera/log/info_20250722-185439.6373
