+++ camera/examples/Camera/readme.txt
