+++ image_processing/ApplyGaborFilter.py
