+++ image_processing/ImageCompressor.py
