+++ image_processing/MaskAreaOutsideBall.py
