+++ image_processing/__init__.py
@@ -1,0 +1,1 @@
+"Image preprocessing utilities."
