+++ scripts/manual_ball_picker.py
