+++ spin/CompareRotationImage.py
