+++ spin/GetBallRotation.spec
