+++ spin/GolfBall.py
