+++ spin/RotationCandidate.py
