+++ spin/RotationSearch.py
