+++ spin/__init__.py
@@ -1,0 +1,1 @@
+"Golf ball spin analysis."
